requirements.txt
Flask
Flask_SQLAlchemy
Flask_Migrate
python-dotenv
psycopg2-binary
Flask-WTF
Flask-Mail
requests
flask-talisman
flask-limiter
cryptography
pyjwt>=2.0.0
cryptography>=3.0
requests pyfiglet rich
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
Flask-WTF
email-validator
python-dotenv
requests
gunicorn
Flask-Cors
Flask-RESTful
Flask-HTTPAuth
Flask-Mail
Flask-Admin
Flask-Security
Flask-Cache
Flask-Session
Flask-Babel
Flask-Assets
Flask-Testing
Flask-Uploads
Flask-SocketIO
Flask-API
Flask-JSON
Flask-RESTPlus
Flask-RESTful-Swagger-3
Flask-User
Flask-Principal
Flask-OpenID
Flask-Login
Flask-User-Agent
Flask-APScheduler
Flask-Blueprint
Flask-HTTPAuth
Flask-JSONRPC
Flask-RESTful-Swagger
Flask-RESTPlus-Swagger
Flask-Cache-Control
Flask-Compress
Flask-Login
Flask-Security-Too
